swipl 9.1.19 compatibility: PL_new_term_refs takes size_t
As title; the prototype of this function changed in 9.1.19 [1]

Not sure if it's acceptable to bump the needed version of swipl like this
or if we want some kind of magic to work with older headers too...

[1] https://swi-prolog.discourse.group/t/ann-swi-prolog-9-1-19/6978

+++ swipl/src/dict.rs
@@ -206,8 +206,7 @@ unsafe impl<'a> TermPutable for DictBuilder<'a> {
 
         let tag_term = context.new_term_ref();
         let len = self.entries.len();
-        // TODO assert len is not too big
-        let value_terms = unsafe { fli::PL_new_term_refs(len as i32) };
+        let value_terms = unsafe { fli::PL_new_term_refs(len) };
         let mut value_term = value_terms;
         let mut key_atoms = Vec::with_capacity(len);
